refactor: package cleanup
