--- ignite/pkg/cosmostestutil/sample/sample.go
@@ -1,13 +1,0 @@
-package cosmostestutilsample
-
-import (
-	"github.com/cosmos/cosmos-sdk/crypto/keys/ed25519"
-	sdk "github.com/cosmos/cosmos-sdk/types"
-)
-
-// AccAddress returns a sample account address.
-func AccAddress() string {
-	pk := ed25519.GenPrivKey().PubKey()
-	addr := pk.Address()
-	return sdk.AccAddress(addr).String()
-}
